Some rework, restructure and clean up
