+++ fad_backend/portal/serializers/ProgramSerializer.py
@@ -1,0 +1,14 @@
+from rest_framework.response import Response
+from rest_framework import status
+from drf_yasg.utils import swagger_auto_schema
+from drf_yasg import openapi
+
+from rest_framework import serializers
+
+from portal.models import User, Program, Group
+
+
+class ProgramSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Program
+        fields = '__all__'  # Or specify fields like ['id', 'name'] if known
