+++ fad_backend/portal/views/fadder_overview_view.py
@@ -8,27 +8,29 @@ from rest_framework import serializers
 
 from portal.models import User, Program, Group
 
-
-class ProgramSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Program
-        fields = '__all__'  # Or specify fields like ['id', 'name'] if known
-        # fields = ['name']
+from portal.serializers.UserModelSerializer import UserModelSerializer
 
 
-class GroupSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Group
-        fields = '__all__'  # Or specify fields like ['id', 'name'] if known
-        # fields = ['name', 'attributes']
+# class ProgramSerializer(serializers.ModelSerializer):
+#     class Meta:
+#         model = Program
+#         fields = '__all__'  # Or specify fields like ['id', 'name'] if known
 
-class UserModelSerializer(serializers.ModelSerializer):
-    program = ProgramSerializer(read_only=True)
-    group = GroupSerializer(read_only=True, allow_null=True)
 
-    class Meta:
-        model = User
-        fields = ['user_id', 'role', 'program', 'group', 'attributes']
+
+# class GroupSerializer(serializers.ModelSerializer):
+#     class Meta:
+#         model = Group
+#         fields = '__all__'  # Or specify fields like ['id', 'name'] if known
+
+
+# class UserModelSerializer(serializers.ModelSerializer):
+#     program = ProgramSerializer(read_only=True)
+#     group = GroupSerializer(read_only=True, allow_null=True)
+
+#     class Meta:
+#         model = User
+#         fields = ['user_id', 'role', 'program', 'group', 'attributes'] # TODO: Add diet, shirt_size etc
 
 class UserListView(APIView):
     @swagger_auto_schema(responses={200: UserModelSerializer(many=True)})
